+++ src/components/layout/TheNavbar.vue
@@ -28,7 +28,7 @@ const route = useRoute()
 
           <algolia-doc-search />
 
-          <a href="https://ggt.ink/PoVKNkO" rel="noreferrer">
+          <a href="https://news.pythoncheatsheet.org/" rel="noreferrer">
             <base-badge class="hidden lg:flex">
               <span class="hidden xl:flex">
                 ✨ Read the latest Python Cheatsheet Issue!
@@ -89,7 +89,11 @@ const route = useRoute()
           </div>
 
           <base-theme-toggle />
-          <a target="_blank" href="https://ggt.ink/vf7kKkS" rel="noreferrer">
+          <a
+            target="_blank"
+            href="https://github.com/wilfredinni/python-cheatsheet"
+            rel="noreferrer"
+          >
             <github-icon />
             <span class="sr-only">python cheatsheet repository</span>
           </a>