+++ src/pages/sponsor.vue
@@ -58,7 +58,7 @@ useHead({
 
     <base-customizable-link-card
       title="Newsletter"
-      path="https://ggt.ink/2CUG8y7"
+      path="https://news.pythoncheatsheet.org/sponsorship"
       :is-external="true"
     >
       Reach an ever-growing number of Python developers.
@@ -72,7 +72,7 @@ useHead({
 
     <base-customizable-link-card
       title="GitHub"
-      path="https://ggt.ink/Q5HjIVK"
+      path="https://github.com/sponsors/wilfredinni"
       :is-external="true"
     >
       If you find this project useful, consider becoming an Sponsor on GitHub.